Eliminated search logging
Search logging means that the customer controls what autosuggests appear.
In most cases, the autosuggests should not be provided by those who are trying
to figure out what they are searching for. It makes more sense for it to come
from the search engine, or from a seeded list provided by the vendor